third_party/WebKit/LayoutTests/external/wpt/mediacapture-image/idlharness-expected.txt

This is a testharness.js-based test.
PASS ImageCapture interface: existence and properties of interface object
PASS ImageCapture interface object length
PASS ImageCapture interface object name
FAIL ImageCapture interface: existence and properties of interface prototype object assert_equals: class string of ImageCapture.prototype expected "[object ImageCapturePrototype]" but got "[object ImageCapture]"
PASS ImageCapture interface: existence and properties of interface prototype object's "constructor" property
FAIL ImageCapture interface: operation takePhoto(PhotoSettings) assert_throws: calling operation with this = null didn't throw TypeError function "function () {
            fn.apply(obj, args);
        }" did not throw
FAIL ImageCapture interface: operation getPhotoCapabilities() assert_throws: calling operation with this = null didn't throw TypeError function "function () {
            fn.apply(obj, args);
        }" did not throw
FAIL ImageCapture interface: operation grabFrame() assert_throws: calling operation with this = null didn't throw TypeError function "function () {
            fn.apply(obj, args);
        }" did not throw
PASS ImageCapture interface: attribute track
PASS PhotoCapabilities interface: existence and properties of interface object
PASS PhotoCapabilities interface object length
PASS PhotoCapabilities interface object name
FAIL PhotoCapabilities interface: existence and properties of interface prototype object assert_equals: class string of PhotoCapabilities.prototype expected "[object PhotoCapabilitiesPrototype]" but got "[object PhotoCapabilities]"
PASS PhotoCapabilities interface: existence and properties of interface prototype object's "constructor" property
PASS PhotoCapabilities interface: attribute redEyeReduction
PASS PhotoCapabilities interface: attribute imageHeight
PASS PhotoCapabilities interface: attribute imageWidth
PASS PhotoCapabilities interface: attribute fillLightMode
PASS MediaSettingsRange interface: existence and properties of interface object
PASS MediaSettingsRange interface object length
PASS MediaSettingsRange interface object name
FAIL MediaSettingsRange interface: existence and properties of interface prototype object assert_equals: class string of MediaSettingsRange.prototype expected "[object MediaSettingsRangePrototype]" but got "[object MediaSettingsRange]"
PASS MediaSettingsRange interface: existence and properties of interface prototype object's "constructor" property
PASS MediaSettingsRange interface: attribute max
PASS MediaSettingsRange interface: attribute min
PASS MediaSettingsRange interface: attribute step
Harness: the test ran to completion.
